fix bug
remove routes to fix unmatched route
