+++ tsconfig.json
@@ -13,5 +13,5 @@
     "**/*.tsx",
     ".expo/types/**/*.ts",
     "expo-env.d.ts",  
-    "app/(tabs)/_layout.jsx", "app/_layout.jsx",]
+  ]
 }
